requirements.txt
pandas>=2.2
pyarrow>=16.0
pyyaml>=6.0
jinja2>=3.1
python-dateutil>=2.9
requests>=2.31
tenacity>=8.2
streamlit>=1.28.0
numpy>=1.26
